Remove abstract class MediumPeriodicFormatter
